correct rdm() that was always returning 0 due to incorrect cast

+++ src/tools.h
@@ -5,5 +5,5 @@ void grid_census(struct RUNPARAMS *param, struct CPUINFO *cpu);
 void checkMtot(struct OCT **firstoct, struct RUNPARAMS *param, struct CPUINFO *cpu);
 void myradixsort(int *a,int n);
 
-REAL rdm(REAL a, REAL b);
+REAL rdm(double a, double b);
 unsigned int gpoiss(REAL lambda);
